+++ lib/src/embeddings/bge.dart
@@ -17,6 +17,7 @@ class BgeEmbedderFactory {
     String? tokenizerFile,
     bool includeExternalData = true,
     DownloadProgress? onProgress,
+    int maxConnections = 1,
     bool force = false,
     String? hfToken,
   }) async {
@@ -29,6 +30,7 @@ class BgeEmbedderFactory {
       tokenizerFile: tokenizerFile,
       includeExternalData: includeExternalData,
       onProgress: onProgress,
+      maxConnections: maxConnections,
       force: force,
     );
     return frb.BgeEmbedder.create(